requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
docopt==0.6.2
idna==3.3
loguru==0.6.0
Pillow==9.2.0
pyTelegramBotAPI==4.5.1
python-barcode==0.14.0
python-Levenshtein==0.12.2
qrcode==7.3.1
requests==2.27.1
urllib3==1.26.9
win32-setctime==1.1.0
yarg==0.1.9
